+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # _seqr_ Changes
 
 ## dev
+* Add data type to family analysed by (REQUIRES DB MIGRATION)
 
 ## 4/11/22
 * Remove pre-built static assets (REQUIRES IMAGE UPDATE)